I made a lot of changes

+++ include/global.berry.h
@@ -46,7 +46,7 @@ struct Berry2
 struct EnigmaBerry
 {
     struct Berry2 berry;
-    u8 itemEffect[BERRY_ITEM_EFFECT_COUNT];
+    u16 itemEffect[BERRY_ITEM_EFFECT_COUNT];
     u8 holdEffect;
     u8 holdEffectParam;
     u32 checksum;
@@ -56,7 +56,7 @@ struct BattleEnigmaBerry
 {
     /*0x00*/ u8 name[BERRY_NAME_LENGTH + 1];
     /*0x07*/ u8 holdEffect;
-    /*0x08*/ u8 itemEffect[BERRY_ITEM_EFFECT_COUNT];
+    /*0x08*/ u16 itemEffect[BERRY_ITEM_EFFECT_COUNT];
     /*0x1A*/ u8 holdEffectParam;
 };
